requirements.txt
python-dateutil~=2.8.2
pyyaml>=6.0.1
spacy~=3.7.2
openai~=0.28.1
backoff~=2.2.1